add remember me check box

+++ src/app/core/services/local-storage.service.ts
@@ -1,19 +1,29 @@
 export class LocalStorageService {
   public get(key: string, defaultValue: any = null): string
   {
-    const result = localStorage.getItem(key);
+    const result = (sessionStorage.getItem(key) !== null) ? sessionStorage.getItem(key) : localStorage.getItem(key);
+
     return result !== null ? result : defaultValue;
   }
 
-  public set(key: string, value: string|number)
+  public set(key: string, value: string|number, keep: boolean = false)
   {
-    localStorage.setItem(key, String(value));
+    if (keep)
+    {
+      localStorage.setItem(key, String(value));
+    }
+    else
+    {
+      sessionStorage.setItem(key, String(value));
+    }
     return this;
   }
 
   public remove(key: string)
   {
+    sessionStorage.removeItem(key);
     localStorage.removeItem(key);
+
     return this;
   }
 }
